data-ingestion/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-datacatalog